chore: move formatCurrency into helpers
