--- libs/application/templates/inao/shared/utils/formatCurrency.ts
@@ -1,2 +1,0 @@
-export const formatCurrency = (answer: string) =>
-  answer.replace(/\B(?=(\d{3})+(?!\d))/g, '.') + ' kr.'
